+++ video/d3d11/d3d11_texture.h
@@ -22,8 +22,8 @@ public:
 
 	
 
-	f32 m_h;
-	f32 m_w;
+	u32 m_h;
+	u32 m_w;
 };
 
 #endif
